refactor: move `SoapEnvelope` to `it.pagopa.transactions.utils.soap` package
