+++ src/main/java/it/pagopa/transactions/utils/soap/Jaxb2SoapEncoder.java
@@ -13,7 +13,6 @@ import org.springframework.util.MimeType;
 import org.springframework.util.MimeTypeUtils;
 import org.w3c.dom.Document;
 
-import it.pagopa.transactions.model.SoapEnvelope;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
