+++ src/test/java/it/pagopa/transactions/client/NodeForPspClientTest.java
@@ -1,7 +1,7 @@
 package it.pagopa.transactions.client;
 
 import it.pagopa.generated.transactions.model.*;
-import it.pagopa.transactions.model.SoapEnvelope;
+import it.pagopa.transactions.utils.soap.SoapEnvelope;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
